requirements.txt
feedparser
jinja2
openai
python-dotenv
pyyaml
